% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plot.R
\name{build_gsea_montage}
\alias{build_gsea_montage}
\title{Build the GSEA montage using the output of `plot_gsea_volcano()`:
Showing the volcano plot on upper right corner
And a montage for selected genesets showing the running enrichment score,
gradient and ranks.}
\usage{
build_gsea_montage(p)
}
\arguments{
\item{p}{a GSEA plot generated by `plot_gsea_volcano()`}
}
\description{
Build the GSEA montage using the output of `plot_gsea_volcano()`:
Showing the volcano plot on upper right corner
And a montage for selected genesets showing the running enrichment score,
gradient and ranks.
}
