% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pie.R
\docType{data}
\name{GeomPie}
\alias{GeomPie}
\title{Draw a piechart for binary proportions.}
\format{
An object of class \code{GeomPie} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomPie
}
\description{
Draw a piechart for binary proportions.
}
\keyword{datasets}
