% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{ggplot_survival}
\alias{ggplot_survival}
\title{Wrapper around `ggplot` and `ggkm` to produce repeatedly used Kaplan-Meier curves}
\usage{
ggplot_survival(
  data,
  time,
  event,
  predictor = NULL,
  hr = TRUE,
  count_npcx = NULL,
  count_npcy = 0.05
)
}
\arguments{
\item{data}{A tibble}

\item{time, event, predictor}{the variable names to produce the Kaplan-Meier plot.}

\item{hr}{Boolean telling whether to show the hazard ratio and p-value.}

\item{count_npcx, count_npcy}{position of the counts. If count_npcx is NULL, will try to spread them horizontally.}
}
\description{
The function add statistical annotations (HR and p-value). Grouping the data will show facets with the annotations.
}
