% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plot.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Prepare the GSEA plot}
\usage{
plot_gsea(gsea, gs_tidy, title = "Gene set", lbl = c("Left", "Right"), palette)
}
\arguments{
\item{gsea, gs_tidy}{The GSEA and selected genesets to be highlighted}

\item{title}{Title of the plot}

\item{lbl}{vector of size 2 with labels to be shown on the gradient bar (DGE contrast)}
}
\description{
Prepare the GSEA plot
}
