% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plot.R
\name{set_panel_size}
\alias{set_panel_size}
\title{Set absolute ggplot panel sizes (adapted from https://stackoverflow.com/a/32583612)}
\usage{
set_panel_size(p = NULL, g = ggplotGrob(p), width = NULL, height = NULL)
}
\arguments{
\item{x}{a ggplot}

\item{filename}{The basename of the output file}
}
\description{
Set absolute ggplot panel sizes (adapted from https://stackoverflow.com/a/32583612)
}
