% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{tidy_coxph}
\alias{tidy_coxph}
\title{Tidy output for `survival::coxph()`}
\usage{
tidy_coxph(
  data,
  time,
  event,
  predictor,
  exponentiate = TRUE,
  cox_zph = TRUE,
  ...
)
}
\arguments{
\item{data}{A tibble}

\item{time, event, predictor}{the variable names to peroform the COX proportional hazard model.}

\item{exponentiate}{Boolean telling whether to exponentiate the hazard ratio.}

\item{cox_zph}{Perform the `surival::cox.zph` test to check the assumption for the Cox regression model fit.}
}
\description{
Wrapper around `survival::coxph()` to provide a tidy tibble using broom.
}
