% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plot.R
\name{write_heatmap}
\alias{write_heatmap}
\title{Save a ComplexHeatmap as a standalone pdf file}
\usage{
write_heatmap(x, filename, ..., width = NA, height = NA)
}
\arguments{
\item{x}{a ComplexHeatmap object}

\item{filename}{The basename of the output file}

\item{width, height}{The settings for the output file}
}
\description{
Save a ComplexHeatmap as a standalone pdf file
}
