% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plot.R
\name{write_plot}
\alias{write_plot}
\title{Save a ggplot as a standalone file using ggsave}
\usage{
write_plot(
  x,
  filename,
  device = grDevices::png,
  width = NA,
  height = NA,
  units = "cm",
  ...
)
}
\arguments{
\item{x}{a ggplot}

\item{filename}{The basename of the output file}
}
\description{
Save a ggplot as a standalone file using ggsave
}
